lost password: include user name in mail subject/body
Users often send a mail to help@lcwo.net, asking for a new
password without specifying their user name. Modify the
mailto links in a way that (if known) the username will
appear in the mail, or at least provide some body text that
will give a hint to adding the user name to the mail.